requirements (4).txt
playwright==1.44.0
python-telegram-bot==21.3
google-generativeai==0.7.2
